test: move rng tests to separate file
